isTeacher to register.html,divs for the options for all question types
Added isTeacher checkbox to register.html. The value is passed through
the controller and shown in console.

According to the types of questions there are different divs, containing
the appropriate radio buttons. Because the form is used for both adding
and editing question, there are two divs of each type of question
accordingly referred to as filled(in edit mode) and empty(in add mode).