isTeacher integrated with DB via person.approved field
is_approved field added to "people" table.
All the students are automatically approved whereas the teachers need to
be approved by an admin.  For this is created teachers-for-approval.html
page, visible only for admins. There are "approve" and "deny" icons for
future development.

Logo removed from previously forgotten pages such as userslist.html.