Enums in BG. Scrollable tables. Form buttons to bottom. Required fields.
QuestionType and QuestionDifficulty enums with bulgarian translations as
values.
The function getRoleValue() in Role model returns the bulgarian
translation of the given role.

Tables made with scrollable body in the following:
lesson/index, question/index, user/teachers-for-approval,
user/userslist, lesson/list.

The buttons "Запази" and "Отказ" are fixed to the bottom of the parent
jumbotron in the following:
lesson/form, question/form, gain-knowledge, profile.

Added titles to required fields in the following: profile,
question/form, lesson/form. Implemented logic for the required fields
with js.


Refactored table with the needed data for users/userslist .

Add/Edit question: fixed logic which was not entirely working before.
When changing question types in edit mode, it was creating new question,
etc.