Added sha256 js function. Hashed passwords. JS notifications.
Refactored register view for the fields to be in two columns.
Check if empty before submitting for the fields in register.html.

When creating a user, the password is being hashed and sent to db.
When changing password, the current password hash is compared to the
password in db, the two fields for the new password are being compared
also. If either of the comparisons is not true, a proper error message
is shown.
During a logging attempt, the provided password is hashed and checked
with the real one in the db. If false -> appropriate error message is
shown.