Added Game mode and Ranking. Code refactored.
The game can be filtered by subject and difficulty. Then, the user can
start the game. For now there are 5 questions per game. After the
completion, a review of the given answers and the correct ones is
presented with the amount of points earned.

Created ranking. It can be overall ranking or separated by subjects.
Added ranking table, model etc.

Implemented student table.
Game table/model -> some fields are replaced with others. Now the table
consists of id, createdAt date and the duration of the game.